# server/openapi.yaml
openapi: 3.0.3
info:
  title: AskFlow API
  description: |
    RESTful API for the AskFlow FAQ & Knowledge Base platform.

    ## Authentication
    Use an API key via the `X-API-Key` header to authenticate requests.
    Create API keys in the AskFlow admin panel under Settings > API Keys.

    ## AI & Chatbot Integration
    AskFlow is designed to integrate with chatbots and AI assistants:

    - **`POST /ask`** — Send a question, get the best matching answers (ideal for chatbots)
    - **`GET /context`** — Get all FAQ data for RAG/knowledge base population
    - **`GET /search`** — Full-text search with relevance ranking

    ### Supported Platforms
    - **Custom chatbots** — Use the /ask endpoint directly
    - **Botpress / Dialogflow / Rasa** — Use /ask as a webhook or custom action
    - **LangChain / LlamaIndex** — Use /context to populate vector stores
    - **Claude / ChatGPT tools** — Use /ask as a tool/function call

    ### Quick Start
    ```bash
    # 1. Get your API key from the admin panel
    # 2. Ask a question
    curl -X POST https://your-domain.com/api/v1/ask \
      -H "X-API-Key: faq_your-api-key" \
      -H "Content-Type: application/json" \
      -d '{"question": "How do I get started?"}'
    ```
  version: 1.0.0
  contact:
    name: AskFlow
servers:
  - url: /api/v1
    description: API v1
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    Category:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        items:
          type: array
          items:
            $ref: '#/components/schemas/FAQItem'
    FAQItem:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
...
    SearchResult:
      type: object
      properties:
        id:
          type: string
        category_id:
          type: string
        question:
          type: string
        answer:
          type: string
        category_title:
          type: string
        product_id:
          type: string
        product_name:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
paths:
  /products:
    get:
      summary: List all products with nested categories and FAQ items
      operationId: getProducts
      responses:
        '200':
          description: List of products with full category tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          description: API key missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                  default: ''
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{id}:
    get:
      summary: Get a single product with nested categories
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product with full category tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated product
    delete:
      summary: Delete a product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '404':
          description: Product not found
...
  /search:
    get:
      summary: Search FAQ items
      operationId: searchFaq
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
...